tried to creat overlap coefficient
Tried to find overlap by 1) making normalized histograms of the matching and non matching distributions. 2)making functions for the histograms 3) using the fuctions that describe the histograms  and integrating with scipy quad function to find overlap.

problems: Integration function stops working for q6_n and is't really right for the other columns either.

Fix: If it were possible to use the pandas.plot.kde() to create a function It would work much better. I don't yet know how to make array from it yet though