Created two way to find overlap with histogram and kde
I added a component to the histograms that made it so it really does only measure the overlap of the two histograms. It finds the minimum of each and integrates to that value. an important note is that I found the hightest number of bins possible before there are any empty bins. If there are empty bins the overlap doesn't really work.

I also did the same thing but used a gaussianKDE that I found. I don't know if that value is better.